Sentinel/pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "The backend for the Crown Point Consulting project."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.39.0",
    "python-dotenv",
    "pydantic[email]>=2.12.5",
    "mailjet>=1.4.1",
    "mailjet-rest>=1.5.1",
    "tenacity>=9.1.2",
    "logging>=0.4.9.6",
]
